CARVAJAL_LINUX/requirements.txt
# CARVAJAL VENTAS - Procesador de Reportes para Odoo 14
# ==============================================================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
yagmail
# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Excel Processing
openpyxl==3.1.2
python-dateutil==2.8.2

# Environment Configuration
python-dotenv==1.0.0

# ID Generation
python-ulid==1.1.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Async
aiofiles==23.2.1

# Scheduling (para procesamiento automático)
apscheduler==3.10.4

# SSH/SFTP (para sincronización Windows → Servidor)
paramiko==3.4.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging & Monitoring
python-json-logger==2.0.7

# Optional: Odoo Integration
# xmlrpc2==1.3.3  # Uncomment if using Odoo XML-RPC

# Type hints
typing-extensions==4.8.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.0
